refactor: use floats for positions in icon

+++ src/gui/icon.py
@@ -4,7 +4,7 @@ from src.gui.element import Element
 
 
 class Icon(Element):
-    def __init__(self, id: str, x: int, y: int, image: pygame.Surface):
+    def __init__(self, id: str, x: float, y: float, image: pygame.Surface):
         self.id = id
         self.image = image
         self.rect = self.image.get_rect(topleft=(x, y))
@@ -14,5 +14,5 @@ class Icon(Element):
     def draw(self, surface: pygame.Surface) -> None:
         surface.blit(self.image, self.rect)
 
-    def update(self, delta_time: float, mouse_position: tuple[int, int]) -> None:
+    def update(self, delta_time: float, mouse_position: tuple[float, float]) -> None:
         pass
